refactor: created file for new component
